💄 style: enhance chat input appearance
